+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ next version
 * Add modeling.get_baselines
 * Add modeling.get_covariate_baselines
 * Add modeling.get_doses
+* Add modeling.list_time_varying_covariates
 
 0.40.0 (2021-12-16)
 -------------------